+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.0] - 2026-09-08
+
+### Fixed
+
+- Enchantment removal now supports filtering by item (such as with `not` filters).
+
+### Changed
+
+- Bumped version number for parity with the 1.20.1 version.
+- `/enchant` now blocks banned enchantments.
+
 ## [2.12.1] - 2026-08-14
 
 ### Fixed